refs 123795: rename file
